% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_forest.R
\name{fit_forest}
\alias{fit_forest}
\title{Fit a classification random forest}
\usage{
fit_forest(data, target, ntree = 500, ...)
}
\arguments{
\item{data}{A data frame.}

\item{target}{Name of the outcome column (must be a factor).}

\item{ntree}{Number of trees.}

\item{...}{Extra arguments passed to \code{\link[randomForest:randomForest]{randomForest::randomForest()}}.}
}
\value{
A \code{randomForest} model.
}
\description{
Fit a classification random forest
}
