% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tree.R
\name{fit_tree}
\alias{fit_tree}
\title{Fit a classification decision tree}
\usage{
fit_tree(data, target, ...)
}
\arguments{
\item{data}{A data frame.}

\item{target}{Name of the outcome column (must be a factor).}

\item{...}{Extra arguments passed to \code{\link[rpart:rpart]{rpart::rpart()}}.}
}
\value{
An \code{rpart} model.
}
\description{
Fit a classification decision tree
}
