% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tree.R
\name{plot_tree}
\alias{plot_tree}
\title{Tracer un arbre de décision}
\usage{
plot_tree(model)
}
\arguments{
\item{model}{Un modèle \code{rpart} issu de \code{\link[=fit_tree]{fit_tree()}}.}
}
\description{
Tracer un arbre de décision
}
