% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Évaluer un modèle sur des données de test}
\usage{
evaluate_model(model, test, target)
}
\arguments{
\item{model}{Un modèle issu de \code{\link[=fit_tree]{fit_tree()}} ou \code{\link[=fit_forest]{fit_forest()}}.}

\item{test}{Data frame de test.}

\item{target}{Nom de la variable cible.}
}
\value{
Classification : matrice de confusion et accuracy.
Régression : RMSE, MAE et R².
}
\description{
Évaluer un modèle sur des données de test
}
