% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tree.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot a decision tree}
\usage{
plot_tree(model)
}
\arguments{
\item{model}{An \code{rpart} model from \code{\link[=fit_tree]{fit_tree()}}.}
}
\description{
Plot a decision tree
}
